+++ src/mol-plugin/commands.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2018-2023 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2018-2024 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author David Sehnal <david.sehnal@gmail.com>
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
@@ -15,6 +15,7 @@ import { Structure, StructureElement } from '../mol-model/structure';
 import { PluginState } from './state';
 import { PluginToast } from './util/toast';
 import { Vec3 } from '../mol-math/linear-algebra';
+import { PluginStateSnapshotManager } from '../mol-plugin-state/manager/snapshots';
 
 export const PluginCommands = {
     State: {
@@ -28,7 +29,7 @@ export const PluginCommands = {
         ToggleVisibility: PluginCommand<{ state: State, ref: StateTransform.Ref }>(),
 
         Snapshots: {
-            Add: PluginCommand<{ key?: string, name?: string, description?: string, params?: PluginState.SnapshotParams }>(),
+            Add: PluginCommand<{ key?: string, name?: string, description?: string, descriptionFormat?: PluginStateSnapshotManager.DescriptionFormat, params?: PluginState.SnapshotParams }>(),
             Replace: PluginCommand<{ id: string, params?: PluginState.SnapshotParams }>(),
             Move: PluginCommand<{ id: string, dir: -1 | 1 }>(),
             Remove: PluginCommand<{ id: string }>(),
@@ -64,11 +65,12 @@ export const PluginCommands = {
         Reset: PluginCommand<{ durationMs?: number, snapshot?: Partial<Camera.Snapshot> }>(),
         SetSnapshot: PluginCommand<{ snapshot: Partial<Camera.Snapshot>, durationMs?: number }>(),
         Focus: PluginCommand<{ center: Vec3, radius: number, durationMs?: number }>(),
+        FocusObject: PluginCommand<PluginState.SnapshotFocusInfo & { durationMs?: number }>(),
         OrientAxes: PluginCommand<{ structures?: Structure[], durationMs?: number }>(),
         ResetAxes: PluginCommand<{ durationMs?: number }>(),
     },
     Canvas3D: {
         SetSettings: PluginCommand<{ settings: Partial<Canvas3DProps> | ((old: Canvas3DProps) => Partial<Canvas3DProps> | void) }>(),
-        ResetSettings: PluginCommand<{ }>()
+        ResetSettings: PluginCommand<{}>()
     }
 };